Fix str exception on setup.py for Python 3
In Python 3, `os.environ` values are already in `str` type. There is no need to wrap them with `str` function again.

When you do `str(s, 'ascii')` on a string it raises an exception:
```
TypeError: decoding str is not supported
```

I have removed the `str()` wrapper to fix this issue.