Set up caching
Caching is done for:
 - pyenv collateral on macos
 - The icu4c installed collateral for all os's

Based on my research, the cache is unique for each OS, Python version,
and set of environment variables. Since the ICU version is defined by
the environment variable, the different installations of ICU versions
should not clash with each other if we install them into the same
directory name.