Refactor .travis-ci.yml and build script
Both files have been refactored to make it easier to extend, which needs
to happen to support windows.

The build script has been broken into functions, and the installation
of pyenv and the display of ICU configuration has been moved
to this script to make the .travis-ci.yml cleaner.

The .travis-ci.yml has removed everything that is now in the build
script, and has also gotten rid of the jobs matrix and now just defines
explictly all jobs to execute.