Add potential support for Windows
This adds commented out jobs for Windows. In the bash script, it adds
building python on windows, fetching and extracting the source code on
windows, and a broken attempt at building the binaries on windows.

Currently, the build executes for 13 minutes then fails at the end.
Should someone want to get this working, most of the infrastructure is
there, they just need to mess with the code that actually builds, and
then potentially the PATH when actually executing pip and running the
tests.