Expand Travis-CI scope (#118)
* Travis-CI now recommends to omit sudo altogether

https://blog.travis-ci.com/2018-11-19-required-linux-infrastructure-migration

* Place ICU versions in a matrix

This is likely WAY too many ICU versions to actually test against,
but it is exhausitve as far as I can tell.

* Test against all currently support Python versions

* Move before_install into the script ci-install-icu4c.bash

It is expected that the installtion instructions for windows and
will be different from *nix, so to keep the config simple the
installation instructions have been moved to a separate script.

Use ./runConfigureICU instead of ./configure

* Update build matrix to limit the total number of jobs

Previously, every ICU would be tested against every Python. That's just
too many jobs.

Now, the latest ICU is always tested, as well as one older version.
This is done against all Python versions.

Additionally, for the latest Python version all ICU versions are tested.

* Add support for macos

The macos job is at the top because it takes a long time to boot and to
queue.

Also needed was:
 - Add icu4c bin location to the PATH for icu-config
 - Add DYLD_LIBRARY_PATH... it's like LD_LIBRARY_PATH, but for macos.

* Drop unneeded six installation

* Set up caching

Caching is done for:
 - pyenv collateral on macos
 - The icu4c installed collateral for all os's

Based on my research, the cache is unique for each OS, Python version,
and set of environment variables. Since the ICU version is defined by
the environment variable, the different installations of ICU versions
should not clash with each other if we install them into the same
directory name.

* List icule as needed library for ICU < 58

* Add configuration information before PyICU build

ICU introspection commands are now invoked so we can be sure
that we are building against what we think we are building

* Omit ICU version 52.2 from CI

The --cxxflags output from icu-config for this version is broken
(it gives an argument incompatible with gcc).

* Refactor .travis-ci.yml and build script

Both files have been refactored to make it easier to extend, which needs
to happen to support windows.

The build script has been broken into functions, and the installation
of pyenv and the display of ICU configuration has been moved
to this script to make the .travis-ci.yml cleaner.

The .travis-ci.yml has removed everything that is now in the build
script, and has also gotten rid of the jobs matrix and now just defines
explictly all jobs to execute.

* Add potential support for Windows

This adds commented out jobs for Windows. In the bash script, it adds
building python on windows, fetching and extracting the source code on
windows, and a broken attempt at building the binaries on windows.

Currently, the build executes for 13 minutes then fails at the end.
Should someone want to get this working, most of the infrastructure is
there, they just need to mess with the code that actually builds, and
then potentially the PATH when actually executing pip and running the
tests.